// src/app/models/renewal/renewal.ts
export class Renewal {
    name?: string;
    surname?: string;
    dateOfBirth?: any;
    sex?: string;
    passportNumber?: string;
    passportIssueDate?: any;
    passportExpiryDate?: any;
    nationality?: string;
    institution?: string;
    course?: string;
    studyYear?: any;
    duration?: any;
    email?: any;
    phone?: any;
    alternatePhone?: any;
    postalAddress?: string;
    residentialAddress?: string;
    status?: string;
    approved?: string;
    registrationProofFile?: any;
    passportFile?: any;
    residenceProofFile?: any;
    criminalRecordNamibiaFile?: any;
    criminalRecordOriginFile?: any;

}